Verificando se dada casa legislativa existe no banco, antes da exclusão.

+++ radar_parlamentar/modelagem/models.py
@@ -305,8 +305,17 @@ class CasaLegislativa(models.Model):
         """Método que deleta determinado registro de casa legislativa em cascata
             Argumentos:
                 nome_casa - Nome da casa a ser deletada"""
-        CasaLegislativa.objects.filter(nome_curto=nome_casa_curto).delete()
+        
+        allCasaLegislativa = CasaLegislativa.objects.all()
+        achei = False
+        for casa in allCasaLegislativa:
+            if nome_casa_curto == casa.nome_curto:        
+                CasaLegislativa.objects.filter(nome_curto=nome_casa_curto).delete()
+                achei = True
+        if achei != True:
+            print 'A Casa que deseja excluir não existe no banco'
 
+                
 
 class PeriodoCasaLegislativa(object):
     """Atributos:
